+++ contracts/deploy/010_upgrade_single_asset_staking.js
@@ -10,6 +10,7 @@ const {
   log,
   deployWithConfirmation,
   executeProposal,
+  sendProposal
 } = require("../utils/deploy");
 const { proposeArgs } = require("../utils/governor");
 const { getTxOpts } = require("../utils/tx");
