+++ dapp/src/constants/merkleProofedAccountsToBeCompensated.json
@@ -1,1 +1,1 @@
-{"0x1974f84881Af4204a21f18c43D7c4d9Dee331Bb5":{"address":"0x1974f84881Af4204a21f18c43D7c4d9Dee331Bb5","eligible_ousd_value_human":"97,215.09","ousd_compensation_human":"25,053.77","ogn_compensation_w_interest_human":"604,568.70","ogn_compensation_human":"483,654.96","eligible_ousd_value":"97215092642050903900160","ousd_compensation":"25053773160512725975040","ogn_compensation":"483654956310577600034316","index":0,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0x55ed40c9d61dd9bd9970496e257452b0780a6bd51ba1ccadc856aaa50700a716","0xa64e3287cd0f81ecfcb33c42d6b48a81c2433b3a404cb09a6333c36a1d74f18b","0xbfd42bee3723d9331bfa39df43d291e0b61a55408e8a874db30cc804fcb256d3"]},"0x8e097ed5FC6B357Ff15a9a7f3D41cDF5B4a05553":{"address":"0x8e097ed5FC6B357Ff15a9a7f3D41cDF5B4a05553","eligible_ousd_value_human":"5,011.17","ousd_compensation_human":"2,002.79","ogn_compensation_w_interest_human":"25,204.27","ogn_compensation_human":"20,163.41","eligible_ousd_value":"5011174949263366946816","ousd_compensation":"2002793737315841736704","ogn_compensation":"20163412948709954491367","index":1,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0x149fe29c40f47b564cf7bc5c160c6b3222508fbe3b62cd1d7ad5759260d540dc","0xa64e3287cd0f81ecfcb33c42d6b48a81c2433b3a404cb09a6333c36a1d74f18b","0xbfd42bee3723d9331bfa39df43d291e0b61a55408e8a874db30cc804fcb256d3"]},"0xFc1850fDd03F596867318EbD303d6256150d657e":{"address":"0xFc1850fDd03F596867318EbD303d6256150d657e","eligible_ousd_value_human":"2,795.91","ousd_compensation_human":"1,448.98","ogn_compensation_w_interest_human":"11,284.63","ogn_compensation_human":"9,027.70","eligible_ousd_value":"2795910845757201580032","ousd_compensation":"1448977711439300395008","ogn_compensation":"9027701972640088371474","index":2,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0xcb59bdf6b08b42dac34046b2c4176e56b4d7db268c960e22188076fd9d0eb13b","0x202b9fe1b75a7064fe6f13473e52e50ea9ccbb478c430356202fe02f67131ada","0xbfd42bee3723d9331bfa39df43d291e0b61a55408e8a874db30cc804fcb256d3"]},"0x17BAd8cbCDeC350958dF0Bfe01E284dd8Fec3fcD":{"address":"0x17BAd8cbCDeC350958dF0Bfe01E284dd8Fec3fcD","eligible_ousd_value_human":"1,245.74","ousd_compensation_human":"1,061.43","ogn_compensation_w_interest_human":"1,544.08","ogn_compensation_human":"1,235.27","eligible_ousd_value":"1245735535819017682944","ousd_compensation":"1061433883954754420736","ogn_compensation":"1235265763165303366005","index":3,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0xcc8d8e2eadd683f6321e121091603627251a411c10a6b028bcc0b7ee1ffda037","0x202b9fe1b75a7064fe6f13473e52e50ea9ccbb478c430356202fe02f67131ada","0xbfd42bee3723d9331bfa39df43d291e0b61a55408e8a874db30cc804fcb256d3"]},"0x3d89e78a9Feb7Be032D5aC01A10Ee2Ca97Ab35FD":{"address":"0x3d89e78a9Feb7Be032D5aC01A10Ee2Ca97Ab35FD","eligible_ousd_value_human":"5,017.45","ousd_compensation_human":"2,004.36","ogn_compensation_w_interest_human":"25,243.72","ogn_compensation_human":"20,194.98","eligible_ousd_value":"5017454359048619032576","ousd_compensation":"2004363589762154758144","ogn_compensation":"20194978346424023287077","index":4,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0x02d531348ea32d7089989dfc4f0c9bd6f61de1acb98fda3893ac9f494212e531","0x4011e19d10156a70390724d0d45fc25e64cc9e3ab3141dfa2e381791aecd5b4a","0x2928df995100c0e2cf189b112d02fa96cfb74feaeca65788ce1fa06a677cca13"]}}
+{"0x1974f84881Af4204a21f18c43D7c4d9Dee331Bb5":{"address":"0x1974f84881Af4204a21f18c43D7c4d9Dee331Bb5","eligible_ousd_value_human":"97,215.09","ousd_compensation_human":"25,053.77","ogn_compensation_w_interest_human":"604,568.70","ogn_compensation_human":"483,654.96","eligible_ousd_value":"97215092642050903900160","ousd_compensation":"25053773160512725975040","ogn_compensation":"483654956310577600034316","index":0,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0x405917298f03c498d0dc329f7b469d6e16fba182d1acbb4953b0204dd31318dc","0x191515f9ac7291b054622f5198a32c82a986bcfa48e214dcee59b5fff307a58e","0xe4802ab8ebc2bae114abc779e1a04ccf925032db3af679796542cfa2fb0f7f69"]},"0x8e097ed5FC6B357Ff15a9a7f3D41cDF5B4a05553":{"address":"0x8e097ed5FC6B357Ff15a9a7f3D41cDF5B4a05553","eligible_ousd_value_human":"5,011.17","ousd_compensation_human":"2,002.79","ogn_compensation_w_interest_human":"25,204.27","ogn_compensation_human":"20,163.41","eligible_ousd_value":"5011174949263366946816","ousd_compensation":"2002793737315841736704","ogn_compensation":"20163412948709954491367","index":1,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0xd79d17c7ba3bfe94bbf5ec56dda1483006cba01a88398401701157a4eb663801","0x191515f9ac7291b054622f5198a32c82a986bcfa48e214dcee59b5fff307a58e","0xe4802ab8ebc2bae114abc779e1a04ccf925032db3af679796542cfa2fb0f7f69"]},"0xFc1850fDd03F596867318EbD303d6256150d657e":{"address":"0xFc1850fDd03F596867318EbD303d6256150d657e","eligible_ousd_value_human":"2,795.91","ousd_compensation_human":"1,448.98","ogn_compensation_w_interest_human":"11,284.63","ogn_compensation_human":"9,027.70","eligible_ousd_value":"2795910845757201580032","ousd_compensation":"1448977711439300395008","ogn_compensation":"9027701972640088371474","index":2,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0x737e6f182e5ca82a9d904c836a6cacace8b9cd89c9f0fb47378a6e8281941a2a","0x11ac47cdf0448b98a3a413c1c3077c29a740f4a660887d4b53774b8255b51250","0xe4802ab8ebc2bae114abc779e1a04ccf925032db3af679796542cfa2fb0f7f69"]},"0x17BAd8cbCDeC350958dF0Bfe01E284dd8Fec3fcD":{"address":"0x17BAd8cbCDeC350958dF0Bfe01E284dd8Fec3fcD","eligible_ousd_value_human":"1,245.74","ousd_compensation_human":"1,061.43","ogn_compensation_w_interest_human":"1,544.08","ogn_compensation_human":"1,235.27","eligible_ousd_value":"1245735535819017682944","ousd_compensation":"1061433883954754420736","ogn_compensation":"1235265763165303366005","index":3,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0xbcf11b6d38a3f97d6db8949711dc960d5efab528b753acba0a6827a15f30596c","0x11ac47cdf0448b98a3a413c1c3077c29a740f4a660887d4b53774b8255b51250","0xe4802ab8ebc2bae114abc779e1a04ccf925032db3af679796542cfa2fb0f7f69"]},"0x3d89e78a9Feb7Be032D5aC01A10Ee2Ca97Ab35FD":{"address":"0x3d89e78a9Feb7Be032D5aC01A10Ee2Ca97Ab35FD","eligible_ousd_value_human":"5,017.45","ousd_compensation_human":"2,004.36","ogn_compensation_w_interest_human":"25,243.72","ogn_compensation_human":"20,194.98","eligible_ousd_value":"5017454359048619032576","ousd_compensation":"2004363589762154758144","ogn_compensation":"20194978346424023287077","index":4,"type":1,"duration":31104000,"rate":"250000000000000000","proof":["0xe55f53f695e595c9bf1ed375314358ab9848d773ee299d24e4a5d952c642d460","0x61d6f95d013ae57f8883993db7e5396d75f65c09465e3362fb44d8451cd473f3","0x5c75deb2b46c581f04d2d9f45bb131588821a3574fb734bd46b58645d0933f8c"]}}
